consolidate scanner user functions
